refactor(accelerator): extract server::probe health-probe from server.rs (Q2 3/n)
The redundant-instance probe (is_healthy_aztec_response + healthy_aztec_on_port —
classifies a lost :59833 bind: incumbent-us vs foreign squatter) + its test move
verbatim into src/server/probe.rs. server.rs gets `mod probe; pub use
probe::healthy_aztec_on_port;` (so server::healthy_aztec_on_port stays the public
path for main.rs); probe.rs reaches the parent's private PORT via super.

Behavior-preserving: 127 lib tests (classifies_health_responses now runs as
server::probe::tests::*) + clippy -D warnings + headless all green.

Q2 continues with handlers (health + prove core + authorize_origin). Builds on
bind (#309) + tls (#310), both merged.
